Merged in siop-editorial (pull request #41)
Revising of introductory text and scope of SIOPv2

* Changes from Kristina's Suggestions
* Revising of introductory text and scope.

A significant amount of changes to the introductory text, scopes, and
use cases.

Part of this effort is to begin to reference specifications like
Verifiable Credentials and DIDs earlier (ideally during the
introduction) while referring to these in OpenID-specific terminology
within normative text.

Approved-by: Jeremie Miller
Approved-by: Kristina Yasuda